cleanup: move specifically aged bushes into their superclasses
